+++ src/index.js
@@ -1,5 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App";
+import App from "./components/App";
 
 ReactDOM.createRoot(document.getElementById("root")).render(<App />);
